Fixed basic result retrieval
The generic data retrieval view was fixed. Some other minor additions
like support for pagination, change in gerrit change ID field type and
addition of another Manifest view.

+++ benchmarks/models.py
@@ -50,8 +50,8 @@ class Result(models.Model):
     revision = models.CharField(max_length=32, null=True, blank=True)
     configuration = models.ForeignKey(Configuration, blank=True, null=True)  # is it needed ?
     timestamp = models.DateTimeField(null=True)
-    gerrit_change_number = models.DecimalField(blank=True, null=True, max_digits=9, decimal_places=2)
-    gerrit_patchset_number = models.DecimalField(blank=True, null=True, max_digits=9, decimal_places=2)
+    gerrit_change_number = models.IntegerField(blank=True, null=True)
+    gerrit_patchset_number = models.IntegerField(blank=True, null=True)
     gerrit_change_url = models.URLField(blank=True, null=True)
     gerrit_change_id = models.CharField(max_length=42, blank=True, null=True)
     build_url = models.URLField(null=True, blank=True)
